refactor(core): extract /percy/comparison/upload handler into comparison-upload.js
Move handleComparisonUpload out of api.js into its own module, mirroring the
handleSyncJob extraction pattern. Promote the shared encodeURLSearchParams
helper to utils.js (alongside the other URL helpers) so both api.js and the
new module can use it without a circular import. No behavior change.
